requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.23.0
slack-bolt==1.18.0
python-dotenv==1.0.0
apscheduler==3.10.4

